+++ src/Util/CoreStatics.h
@@ -3,7 +3,7 @@
 namespace CoreStatics
 {
 	constexpr static bool IsDebugBuild = _DEBUG;
-	constexpr static double OneMillisec = 1.0 / 1000.0;
+	constexpr static float OneMillisec = 1.0f / 1000.0f;
 	constexpr static unsigned int MaxNumComponents = 32;
 	constexpr static unsigned int MaxNumEntities = -1;
 } // ~namespace CoreStatics
